Restore app routes file
* Something is coupled to formik in a way that breaks errors when the routes
are moved
