+++ src/routes.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
-import AppContainer from 'components/app-container';
+import AppContainer from './components/app-container';
 import Route404 from 'components/404-route';
 import NoMatch from 'components/no-match';
 import EligiblePage from 'pages/eligible';
@@ -10,6 +10,51 @@ import wizardRouteConfig from 'route-config';
 import FSMRouter, { MachineConsumer, MachineState } from 'components/fsm';
 import Wizard from 'components/wizard';
 import ApplicantReview from 'components/review';
+import WorkerSearch from 'components/worker-review/search';
+import WorkerReview from 'components/worker-review/review';
+import workerConfig from 'state-charts/worker';
+
+class Test extends React.Component {
+  render() {
+    return (
+      <FSMRouter config={workerConfig} usePath>
+        <MachineConsumer>
+          {(transition) => {
+            return (
+              <MachineState>
+                {(state) => {
+                  const providedProps = {
+                    transition,
+                    machineState: state
+                  };
+
+                  return (  
+                    <AppContainer>
+                      <Switch>
+                        <Route
+                          path="/worker/search"
+                          render={() =>
+                            <WorkerSearch {...providedProps} />  
+                          }
+                        />
+                        <Route
+                          path="/worker/review"
+                          render={() =>
+                            <WorkerReview {...providedProps} />
+                          }
+                        />
+                      </Switch>
+                    </AppContainer>
+                  );
+                }}
+              </MachineState>
+            );
+          }}
+        </MachineConsumer>
+      </FSMRouter>
+    )
+  }
+}
 
 const ClientRoutes = () =>
   <FSMRouter config={fsmConfig}>
@@ -61,4 +106,11 @@ const ClientRoutes = () =>
     </MachineConsumer>
   </FSMRouter>;
 
-  export default ClientRoutes;
+const Routes = () => (
+  <Switch>
+    <Route path="/worker" component={Test} />
+    <Route path="/" render={() => <ClientRoutes />} />
+  </Switch>
+);
+
+export default Routes;
